Spectral Moments Praat Script
This Praat script calculates the spectral moments of center of gravity, standard deviation, skewness, and kurtosis to segments at 10% intervals as well as to the 20% mid-point values. It also applies a Hann band pass filter prior to calculating the moments. The user can also specify the power for the moments (typically 1, 2, or 2/3). The script runs on all wav files with textgrids that have an interval tier in the specified folder.